+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 - Add context to all parse functions
 - Remove `#[repr(...)]`, instead specify repr in `#[protocol(discriminant = "...")]`
 - Remove Hints
-- Add `#[protocol(auto)]` for automatically calculating element count on write
+- Add `#[protocol(value = "<expr>")]` for automatically writing arbitrary element value